Alteração de controllers para utilizar métodos de services

+++ src/Service/SubTaskService.cs
@@ -1,9 +1,17 @@
-﻿using Domain.Models;
+﻿using Domain.Mappers;
+using Domain.Models;
 using Infra.Repositories;
 
 namespace Service;
 
-public class SubTaskService
+public interface ISubTaskService
+{
+    SubTask Create(SubTaskRequest subTaskRequest);
+    List<SubTask> List(int mainTaskId);
+    SubTask Update(SubTaskUpdate subTaskUpdate, int subTaskId);
+    void Delete(int subTaskId);
+}
+public class SubTaskService : ISubTaskService
 {
     private readonly ISubTaskRepository _subTaskRepository;
 
@@ -12,24 +20,36 @@ public class SubTaskService
         _subTaskRepository = subTaskRepository;
     }
 
-    public SubTask Create(SubTask subTask)
+    public SubTask Create(SubTaskRequest subTaskRequest)
     {
-        return _subTaskRepository.Create(subTask);
+        var newSubTask = SubTaskMapper.ToClass(subTaskRequest);
+        return _subTaskRepository.Create(newSubTask);
     }
 
     public void Delete(int subTaskId)
     {
+        var subTask = _subTaskRepository.Find(subTaskId);
+
+        if (subTask is null)
+            throw new Exception("subTask not found!");
+
         _subTaskRepository.Delete(subTaskId);
     }
 
     public List<SubTask> List(int mainTaskId)
     {
-        var subTasks = _subTaskRepository.Get(mainTaskId);
-        return subTasks;
+        return _subTaskRepository.Get(mainTaskId);
     }
 
-    public SubTask Update(SubTask subTaskUpdate, int subTaskId)
+    public SubTask Update(SubTaskUpdate updateSubTaskRequest, int subTaskId)
     {
-        return _subTaskRepository.Update(subTaskUpdate, subTaskId);
+        var subTask = _subTaskRepository.Find(subTaskId);
+
+        if (subTask is null)
+            throw new Exception("SubTask not found!");
+
+        subTask.Description = updateSubTaskRequest.Description;
+        subTask.Finished = updateSubTaskRequest.Finished;
+        return _subTaskRepository.Update(subTask);
     }
 }
