+++ src/Service/UserService.cs
@@ -1,8 +1,8 @@
-﻿using Infra.Repositories;
-using Domain.Mappers;
+﻿using Domain.Mappers;
+using Domain.Models;
 using Domain.Request;
 using Domain.Responses;
-using Domain.Models;
+using Infra.Repositories;
 
 namespace Service;
 
@@ -12,7 +12,7 @@ public interface IUserService
     void Delete(int userId);
     User? GetById(int userId);
     List<UserResponse> List();
-    UserResponse Update(UpdatedUserRequest userUpdate);
+    UserResponse Update(UserUpdate userUpdate);
 }
 
 public class UserService : IUserService
@@ -36,6 +36,11 @@ public class UserService : IUserService
 
     public void Delete(int userId)
     {
+        var user = _userRepository.Get(userId);
+
+        if (user is null)
+            throw new Exception("User not found!");
+
         _userRepository.Delete(userId);
     }
 
@@ -51,15 +56,17 @@ public class UserService : IUserService
         return userResponse;
     }
 
-    public UserResponse Update(UpdatedUserRequest userUpdate)
+    public UserResponse Update(UserUpdate userUpdate)
     {
-        var existingUser = _userRepository.Get(userUpdate.Id);
+        var user = _userRepository.Get(userUpdate.Id);
 
-        if (existingUser is null)
+        if (user is null)
             throw new Exception("User not found!");
 
-        var updatedUser = UserMapper.ToEntity(userUpdate);
-        _userRepository.Update(updatedUser);
-        return UserMapper.ToResponse(updatedUser);
+        user.Name = userUpdate.Name;
+        user.Email = userUpdate.Email;
+        _userRepository.Update(user);
+        return UserMapper.ToResponse(user);
     }
+    
 }
